FEAT: E 스킬 추가
E 스킬 추가 (지금은 아무 조건 없이 사용 가능)

스테이지 3 매니저 스크립트 조정 (깃허브 기록 확인해주세요)

+++ Assets/Script/Character/CHAR_CharacterStatus.cs
@@ -6,7 +6,7 @@ public class CHAR_CharacterStatus : MonoBehaviour {
 
     //---------- 공격 장에서 사용한다. ----------
     // 체력.
-    public int HP = 100;
+    public float HP = 100;
     public int MaxHP = 100;
     public float stamina = 100;
     // 공격력.
